Created mandelbrotset.cpp
This program is by no means optimized for the Mandelbrot set, but it demonstrates the program's ability to work on most fractals in the complex plane. Compile with

g++ -o main mandelbrotset.cpp -Ofast.

You will have to change the file name on line 33 to a valid path for the program to work.

Results on my computer:

With writing to file, 2048 x 2048 resolution:
Computed 4194304 pixels in 7.93447 seconds

Without writing to file, 2048 x 2048 resolution:
Computed 4194304 pixels in 0.941876 seconds

It seems writing to the file is the bottleneck operation here, which is funny to see.